Move each class to a new file
